Add files via upload
This is owned in part by the parties that wish to remain anonymous that helped in its conceptual synthesis. It will be released by us when we have reason to believe someone would like us to help anyone. Otherwise, lets protect this fine tuned solution at the direction of the head of anonymous. I do not require funds for this project tendril and would like to keep some of the portions private and protected. 

We believe anon has spoken and if you make it in, it is not my choice but theirs. Consider the topic broached and the original solution 1 design yours with an Apache license. Thnx. Max hax wax, make sure to get stax and bax or we might fax the crax to shax.